+++ lib/libmd/md4c.c
@@ -23,12 +23,13 @@
    documentation and/or software.
  */
 
-#include "md4.h"
+#include <sys/types.h>
 #include <string.h>
+#include "md4.h"
 
 typedef unsigned char *POINTER;
-typedef unsigned short int UINT2;
-typedef unsigned long int UINT4;
+typedef u_int16_t UINT2;
+typedef u_int32_t UINT4;
 
 #define PROTO_LIST(list) list
